Restart audioserver during f/w restart due to encryption
In OMR1, the audioserver was under "class main" in audioserver.rc.
In P, the audioserver is under "class core" in audioserver.rc.
For Full-Disk encryption devices, during the framework restart, under
trigger_restart_framework, only "class_start main" is executed.
Hence the audioserver is not restarted and it is out of sync with
the AudioService.java which is restarted.
Explicitly restarting the audioserver for only the case when framework
restart is due to encryption.

Change-Id: Id8ece641e6bfd856c2e7806315c43c076441142e
Reviewed-on: https://gerrit.mot.com/1251557
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Chitra Rao <chitrasr@motorola.com>
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key
Reviewed-by: Vitor Queiroz Albuquerque <vitoralb@motorola.com>